Updates actions and README for branching changes
We have renamed the the `master` branch to `main` and will use this branch for stable releases only.
We have created a new `develop` branch that is now set as the default. Development PR should be made to this branch.

Updated the actions and README to reflect and document the change.